\name{mlcompare-package}
\alias{mlcompare-package}
\alias{mlcompare}
\docType{package}
\title{
\packageTitle{mlcompare}
}
\description{
\packageDescription{mlcompare}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mlcompare}
\packageIndices{mlcompare}
~~ An overview of how to use the package, including ~~
~~ the most important functions ~~
}
\author{
\packageAuthor{mlcompare}

Maintainer: \packageMaintainer{mlcompare}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
